backend/requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
gunicorn
werkzeug
itsdangerous>=2.1
qrcode[pil]>=7.4
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
python-dotenv
